+++ include/expr/stack.hpp
@@ -13,11 +13,11 @@ enum class Op { Const, Var, Add, Sub, Mul, Div, Neg, Sqrt, Sin, Cos, Exp, Log, P
 struct Node {
   Op op;
   union {
-    double value;
+    FloatType value;
     size_t var_idx_or_exponent;
   };
 
-  Node(Op op, double val) : op(op), value(val) {}
+  Node(Op op, FloatType val) : op(op), value(val) {}
   Node(Op op, size_t idx_or_exp) : op(op), var_idx_or_exponent(idx_or_exp) {}
   Node(Op op) : op(op), value(0) {}
 };
@@ -25,8 +25,8 @@ struct Node {
 using Program = std::vector<Node>;
 
 /// Evaluates the program using a stack-based interpreter
-inline double eval(const Program &prog, const std::vector<double> &vars) {
-  std::vector<double> stack;
+inline FloatType eval(const Program &prog, const std::vector<FloatType> &vars) {
+  std::vector<FloatType> stack;
   stack.reserve(prog.size());
 
   for (const auto &node : prog) {
@@ -34,75 +34,75 @@ inline double eval(const Program &prog, const std::vector<double> &vars) {
     case Op::Const: stack.push_back(node.value); break;
     case Op::Var: stack.push_back(vars[node.var_idx_or_exponent]); break;
     case Op::Add: {
-      double b = stack.back();
+      FloatType b = stack.back();
       stack.pop_back();
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(a + b);
       break;
     }
     case Op::Sub: {
-      double b = stack.back();
+      FloatType b = stack.back();
       stack.pop_back();
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(a - b);
       break;
     }
     case Op::Mul: {
-      double b = stack.back();
+      FloatType b = stack.back();
       stack.pop_back();
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(a * b);
       break;
     }
     case Op::Div: {
-      double b = stack.back();
+      FloatType b = stack.back();
       stack.pop_back();
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(a / b);
       break;
     }
     case Op::Neg: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(-a);
       break;
     }
     case Op::Sqrt: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(std::sqrt(a));
       break;
     }
     case Op::Sin: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(std::sin(a));
       break;
     }
     case Op::Cos: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(std::cos(a));
       break;
     }
     case Op::Exp: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(std::exp(a));
       break;
     }
     case Op::Log: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(std::log(a));
       break;
     }
     case Op::PowInt: {
-      double a = stack.back();
+      FloatType a = stack.back();
       stack.pop_back();
       stack.push_back(intPow(a, node.var_idx_or_exponent));
       break;
@@ -121,7 +121,7 @@ inline double eval(const Program &prog, const std::vector<double> &vars) {
 // enum class Op { Add, Sub, Mul, Div, Neg, Sqrt, Sin, Cos, Exp, Log, PowInt };
 
 // struct Const {
-//   double value;
+//   FloatType value;
 // };
 // struct Var {
 //   size_t index;
@@ -137,15 +137,15 @@ inline double eval(const Program &prog, const std::vector<double> &vars) {
 
 // using Program = std::vector<Node>;
 
-// inline double int_pow(double x, int n) {
+// inline FloatType int_pow(FloatType x, int n) {
 //   if (n == 0) return 1.0;
 //   if (n < 0) return 1.0 / int_pow(x, -n);
 //   if (n & 1) return x * int_pow(x * x, n >> 1);
 //   return int_pow(x * x, n >> 1);
 // }
 
-// inline double eval(const Program &prog, const std::vector<double> &vars) {
-//   std::vector<double> stack;
+// inline FloatType eval(const Program &prog, const std::vector<FloatType> &vars) {
+//   std::vector<FloatType> stack;
 //   stack.reserve(prog.size());
 
 //   for (const auto &node : prog) {
@@ -157,75 +157,75 @@ inline double eval(const Program &prog, const std::vector<double> &vars) {
 //           } else if constexpr (std::is_same_v<T, Var>) {
 //             stack.push_back(vars[arg.index]);
 //           } else if constexpr (std::is_same_v<T, PowConst>) {
-//             double a = stack.back();
+//             FloatType a = stack.back();
 //             stack.pop_back();
 //             stack.push_back(int_pow(a, arg.exponent));
 //           } else if constexpr (std::is_same_v<T, RawOp>) {
 //             switch (arg.op) {
 //             case Op::Add: {
-//               double b = stack.back();
+//               FloatType b = stack.back();
 //               stack.pop_back();
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(a + b);
 //               break;
 //             }
 //             case Op::Sub: {
-//               double b = stack.back();
+//               FloatType b = stack.back();
 //               stack.pop_back();
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(a - b);
 //               break;
 //             }
 //             case Op::Mul: {
-//               double b = stack.back();
+//               FloatType b = stack.back();
 //               stack.pop_back();
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(a * b);
 //               break;
 //             }
 //             case Op::Div: {
-//               double b = stack.back();
+//               FloatType b = stack.back();
 //               stack.pop_back();
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(a / b);
 //               break;
 //             }
 //             case Op::Neg: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(-a);
 //               break;
 //             }
 //             case Op::Sqrt: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(std::sqrt(a));
 //               break;
 //             }
 //             case Op::Sin: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(std::sin(a));
 //               break;
 //             }
 //             case Op::Cos: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(std::cos(a));
 //               break;
 //             }
 //             case Op::Exp: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(std::exp(a));
 //               break;
 //             }
 //             case Op::Log: {
-//               double a = stack.back();
+//               FloatType a = stack.back();
 //               stack.pop_back();
 //               stack.push_back(std::log(a));
 //               break;
